refactor(Product): change id types to Long and convert mapper to interface
- Change primitive int to Long for id fields in ProductDto to match database schema
- Convert ProductMapper from class to interface to leverage MapStruct's code generation
- Add REST endpoints in ProductController for product retrieval

+++ src/main/java/com/codewithmosh/store/dtos/ProductDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @AllArgsConstructor
 @Data
 public class ProductDto {
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -14,7 +14,7 @@ public class ProductDto {
 
     private String price;
 
-    private int categoryId;
+    private Long categoryId;
